backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
sqlalchemy==2.0.23
sqlite3-utils==3.35.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.0
python-dateutil==2.8.2
